chore: move pkg/parser under drivers/s3/internal/pkg
