+++ drivers/s3/internal/pkg/parser/json.go
