+++ drivers/s3/internal/pkg/parser/parquet.go
