+++ drivers/s3/internal/s3.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"github.com/datazip-inc/olake/drivers/abstract"
+	"github.com/datazip-inc/olake/drivers/s3/internal/pkg/parser"
 	"github.com/datazip-inc/olake/pkg/objstorage"
-	"github.com/datazip-inc/olake/pkg/parser"
 	"github.com/datazip-inc/olake/types"
 	"github.com/datazip-inc/olake/utils"
 	"github.com/datazip-inc/olake/utils/logger"
